requirements.txt
fastapi
uvicorn
pokebase==1.3.0
googletrans==4.0.0rc1
google-generativeai
